requirements.txt
discord.py
psycopg2-binary
pixivpy
pybooru
praw
requests
